agentcore-research/agent/myresearchagent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "myresearchagent"
version = "0.1.0"
description = "AgentCore Runtime Application using Strands SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws-opentelemetry-distro",
    "bedrock-agentcore >= 1.0.3",
    "botocore[crt] >= 1.35.0",
    "strands-agents >= 1.13.0",
    "strands-agents-tools >= 0.1.0",
    "playwright",
    "nest-asyncio",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
